Remove workspace tests and replace with standalone
