+++ package.json
@@ -149,7 +149,6 @@
   },
   "devDependencies": {
     "@types/yargs": "^17.0.10",
-    "cypress": "^11.2.0",
     "http-server": "^14.1.1",
     "jest-dev-server": "^6.1.1"
   },
