+++ tests_fixtures/.gitignore
@@ -1,1 +1,1 @@
-workspace_test/
+standalone_test/
