+++ tests_fixtures/standalone_template/.eslintrc.js
