+++ tests_fixtures/standalone_template/.gitignore
