+++ tests_fixtures/standalone_template/.prettierrc.js
