+++ tests_fixtures/standalone_template/.yarn/releases/yarn-3.2.2.cjs
