+++ tests_fixtures/standalone_template/.yo-rc.json
