+++ tests_fixtures/standalone_template/LICENSE
