+++ tests_fixtures/standalone_template/cypress/.gitignore
