+++ tests_fixtures/standalone_template/cypress/e2e/health-check.cy.ts
