+++ tests_fixtures/standalone_template/cypress/fixtures/example.json
