+++ tests_fixtures/standalone_template/cypress/support/commands.ts
