+++ tests_fixtures/standalone_template/cypress/support/e2e.ts
