+++ tests_fixtures/standalone_template/cypress/tsconfig.json
