+++ tests_fixtures/standalone_template/jest.config.js
