+++ tests_fixtures/standalone_template/src/app/MainApp.tsx
