+++ tests_fixtures/standalone_template/src/app/index.ts
