+++ tests_fixtures/standalone_template/src/assets/datavisyn_favicon.png
