+++ tests_fixtures/standalone_template/src/index.initialize.tsx
