+++ tests_fixtures/standalone_template/src/index.template.ejs
@@ -12,8 +12,6 @@
 </head>
 
 <body>
-  <%=require('tdp_core/dist/template/_body.html')%>
-
   <div id="main" style="height: 100vh;width: 100vw;display: flex;flex-direction: column;"></div>
 </body>
