+++ tests_fixtures/standalone_template/src/phovea.ts
@@ -1,0 +1,2 @@
+// import type { IRegistry } from 'tdp_core';
+// export default (registry: IRegistry) => {};
