+++ tests_fixtures/standalone_template/src/phovea_registry.ts
