+++ tests_fixtures/standalone_template/src/scss/main.scss
