+++ tests_fixtures/standalone_template/tests/index.test.ts
