+++ tests_fixtures/standalone_template/tsconfig.eslint.json
