+++ tests_fixtures/standalone_template/tsconfig.json
