+++ tests_fixtures/standalone_template/tsd.d.ts
