+++ tests_fixtures/standalone_template/typedoc.js
