+++ tests_fixtures/standalone_template/workspace.scss
@@ -7,20 +7,20 @@
 @import '~tdp_core/dist/scss/vendors/bootstrap/1_variables_mixins_utilities';
 
 // map-merge my theme-colors
-@import '~demo/dist/scss/vendors/bootstrap_modify_maps';
+@import './src/scss/vendors/bootstrap_modify_maps';
 
 // Include all Bootstrap components as needed
 @import '~tdp_core/dist/scss/vendors/bootstrap/2_components';
 
 //default app
-@import '~demo/dist/scss/abstracts/variables';
+@import './src/scss/abstracts/variables';
 //dependency tree up
 @import '~tdp_core/dist/scss/abstracts/variables';
 @import '~tdp_core/dist/scss/fontawesome';
 // import mixins
 @import '~tdp_core/dist/scss/abstracts/mixins';
-@import '~demo/dist/scss/abstracts/mixins';
+@import './src/scss/abstracts/mixins';
 //main scss dependency tree down
 @import '~tdp_core/dist/scss/main';
-@import '~demo/dist/scss/main';
+@import './src/scss/main';
 */
