--- tests_fixtures/workspace_template/.gitignore
@@ -1,11 +1,0 @@
-node_modules/
-bundles/
-
-# yarn
-.pnp.*
-.yarn/*
-!.yarn/patches
-!.yarn/plugins
-!.yarn/releases
-!.yarn/sdks
-!.yarn/versions
