--- tests_fixtures/workspace_template/.yo-rc-workspace.json
@@ -1,3 +1,0 @@
-{
-  "defaultApp": "demo"
-}
