--- tests_fixtures/workspace_template/demo/src/phovea.ts
@@ -1,3 +1,0 @@
-import type { IRegistry } from 'tdp_core';
-
-export default (registry: IRegistry) => {};
