--- tests_fixtures/workspace_template/metaData.json
@@ -1,9 +1,0 @@
-{
-  "buildId": "",
-  "name": "",
-  "version": "",
-  "repository": "",
-  "homepage": "",
-  "description": "",
-  "screenshot": ""
-}
