--- tests_fixtures/workspace_template/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "demo_workspace",
-  "version": "0.0.1-SNAPSHOT",
-  "private": true,
-  "author": {
-    "name": "datavisyn GmbH",
-    "email": "contact@datavisyn.io",
-    "url": "https://www.datavisyn.io"
-  },
-  "scripts": {
-    "build:demo": "cd demo && npm run build",
-    "copy:demo": "cd demo && npm run copy",
-    "cy:open:demo": "cd demo && npm run cy:open",
-    "cy:run:demo": "cd demo && npm run cy:run",
-    "delete-dependencies:demo": "cd demo && npm run delete-dependencies",
-    "postinstall": "npm-run-all --parallel 'delete-dependencies:*'",
-    "start": "visyn_scripts start",
-    "build": "npm run webpack:prod",
-    "webpack:dev": "visyn_scripts build --mode development",
-    "webpack:prod": "visyn_scripts build --mode production"
-  },
-  "dependencies": {
-    "demo": "portal:./demo"
-  },
-  "devDependencies": {
-    "npm-run-all": "^4.1.5",
-    "visyn_scripts": "portal:../../"
-  }
-}
