--- tests_fixtures/workspace_template/phovea_registry.ts
@@ -1,1 +1,0 @@
-import 'demo/src/phovea_registry';
