--- tests_fixtures/workspace_template/tailwind.config.js
@@ -1,14 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    './src/**/*.{ts,tsx}',
-  ],
-  theme: {
-    extend: {
-
-    },
-  },
-  plugins: [
-
-  ],
-};
